Attendance pipeline/requirements.txt.txt
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
insightface>=0.7.3
onnxruntime-gpu>=1.16.3   
mediapipe>=0.10.3
ultralytics>=8.1.0
scipy>=1.10.0


# PyTorch with CUDA first
pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# Then all other deps
pip install -r requirements.txt